Move scan state into scan actor
